+++ qa-test-app/src/main/java/com/mapbox/navigation/qa_test_app/view/AppLifecycleActivity.kt
@@ -1,29 +1,24 @@
 package com.mapbox.navigation.qa_test_app.view
 
-import android.annotation.SuppressLint
-import android.os.Build
 import android.os.Bundle
-import android.os.VibrationEffect
-import android.os.Vibrator
-import android.view.View
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.commit
 import androidx.fragment.app.replace
-import com.mapbox.maps.plugin.gestures.gestures
 import com.mapbox.navigation.base.ExperimentalPreviewMapboxNavigationAPI
 import com.mapbox.navigation.base.options.NavigationOptions
 import com.mapbox.navigation.core.lifecycle.MapboxNavigationApp
-import com.mapbox.navigation.core.trip.session.TripSessionState
+import com.mapbox.navigation.dropin.lifecycle.attachResumed
 import com.mapbox.navigation.qa_test_app.R
 import com.mapbox.navigation.qa_test_app.databinding.AppLifecycleActivityLayoutBinding
-import com.mapbox.navigation.qa_test_app.lifecycle.DropInCameraMode
 import com.mapbox.navigation.qa_test_app.lifecycle.DropInContinuousRoutes
-import com.mapbox.navigation.qa_test_app.lifecycle.DropInLocationComponent
-import com.mapbox.navigation.qa_test_app.lifecycle.DropInLocationViewModel
+import com.mapbox.navigation.qa_test_app.lifecycle.DropInLocationPuck
 import com.mapbox.navigation.qa_test_app.lifecycle.DropInNavigationCamera
-import com.mapbox.navigation.qa_test_app.lifecycle.DropInReplayViewModel
 import com.mapbox.navigation.qa_test_app.lifecycle.DropInRoutesInteractor
+import com.mapbox.navigation.qa_test_app.lifecycle.DropInStartReplayButton
+import com.mapbox.navigation.qa_test_app.lifecycle.viewmodel.DropInLocationViewModel
+import com.mapbox.navigation.qa_test_app.lifecycle.viewmodel.DropInNavigationCameraViewModel
+import com.mapbox.navigation.qa_test_app.lifecycle.viewmodel.DropInReplayViewModel
 import com.mapbox.navigation.qa_test_app.utils.Utils.getMapboxAccessToken
 import com.mapbox.navigation.ui.maps.NavigationStyles
 
@@ -35,6 +30,7 @@ class AppLifecycleActivity : AppCompatActivity() {
     }
 
     private val replayViewModel: DropInReplayViewModel by viewModels()
+    private val cameraViewModel: DropInNavigationCameraViewModel by viewModels()
     private val locationViewModel: DropInLocationViewModel by viewModels()
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -61,53 +57,26 @@ class AppLifecycleActivity : AppCompatActivity() {
         // But, they are needed before any MapboxNavigationObserver will call onAttached.
         // Consider moving this to the setup to your Application.onCreate, that way
         // MapboxNavigation can be created whenever you attach a LifecycleOwner.
-        MapboxNavigationApp.setup(
-            NavigationOptions.Builder(this)
-                .accessToken(getMapboxAccessToken(this))
-                .build()
-        )
-
-        // This is commented out to show that MapboxNavigationApp.attach will work when using a
-        // fragment as well. Look into the `RetainedActiveGuidanceFragment` to see it also attaches
-        // itself. In real scenarios, you should attach both LifecycleOwners for better code
-        // maintainability!
-        // MapboxNavigationApp.attach(this)
-
-        // Setup location. The locationViewModel is separate so it can be shared between
-        // the camera, location puck, and the long click listener. The LocationViewModel
-        // will also survive orientation changes.
-        lifecycle.addObserver(DropInLocationComponent(binding.mapView, locationViewModel))
-
-        // Setup the navigation camera.
-        lifecycle.addObserver(
-            DropInNavigationCamera(
-                binding.mapView,
-                DropInCameraMode.FOLLOWING,
-                locationViewModel
+        if (!MapboxNavigationApp.isSetup()) {
+            MapboxNavigationApp.setup(
+                NavigationOptions.Builder(this)
+                    .accessToken(getMapboxAccessToken(this))
+                    .build()
             )
+        }
+
+        /**
+         * Idea for interaction customization.
+         */
+        cameraViewModel.triggerIdleCameraOnMoveListener = false
+
+        attachResumed(
+            DropInLocationPuck(locationViewModel, binding.mapView),
+            DropInRoutesInteractor(locationViewModel, binding.mapView),
+            DropInNavigationCamera(cameraViewModel, locationViewModel, this, binding.mapView),
+            DropInContinuousRoutes(),
+            DropInStartReplayButton(replayViewModel, binding.startNavigation),
         )
-
-        // Setup routes, notice it is not a ViewModel because the state is held by
-        // MapboxNavigation. Here we are re-attaching view state whenever an Activity is created.
-        val routesInteractor = DropInRoutesInteractor(binding.mapView)
-        lifecycle.addObserver(routesInteractor)
-
-        // Setup a long press to find a route.
-        binding.mapView.gestures.addOnMapLongClickListener { point ->
-            vibrate()
-            val originLocation = locationViewModel.locationLiveData.value
-            routesInteractor.findRoute(originLocation, point)
-            false
-        }
-
-        // Setup the map press to select alternative routes.
-        binding.mapView.gestures.addOnMapClickListener { point ->
-            routesInteractor.selectRoute(point)
-            false
-        }
-
-        // Setup continuous alternative routes.
-        lifecycle.addObserver(DropInContinuousRoutes())
 
         // Add active guidance banner with maneuverView and trip progress in a Fragment.
         // This example shows a retained fragment only because they are more complicated and
@@ -119,26 +88,6 @@ class AppLifecycleActivity : AppCompatActivity() {
                 setReorderingAllowed(true)
                 replace<RetainedActiveGuidanceFragment>(R.id.activeGuidanceBannerFragment)
             }
-        }
-
-        // Setup the navigation button.
-        if (TripSessionState.STARTED != MapboxNavigationApp.current()?.getTripSessionState()) {
-            binding.startNavigation.setOnClickListener {
-                binding.startNavigation.visibility = View.GONE
-                replayViewModel.startSimulation()
-            }
-        } else {
-            binding.startNavigation.visibility = View.GONE
-        }
-    }
-
-    @SuppressLint("MissingPermission")
-    private fun vibrate() {
-        val vibrator = getSystemService(VIBRATOR_SERVICE) as Vibrator
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-            vibrator.vibrate(VibrationEffect.createOneShot(100L, VibrationEffect.DEFAULT_AMPLITUDE))
-        } else {
-            vibrator.vibrate(100L)
         }
     }
 }