+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 Mapbox welcomes participation and contributions from everyone.
 
 ## Unreleased
+- Add `MapboxNavigationApp.isSetup` to ensure views do not reset `MapboxNavigation`. Add `MapboxNavigationApp.getObserver` to be able to access registered observers. [#5358](https://github.com/mapbox/mapbox-navigation-android/pull/5358)
+
 #### Features
 
 #### Bug fixes and improvements